notes/transparent mod
